[fixed] duplicate tasks for shuttles causing issue with non-default shuttle times
While loading static objects two tasks for each shuttleport (probably startport aswell? didn't check) are being generated.
With the default values for shuttleports that doesn't seem to cause any issue. If you lower those, you'll have an overlapping departure/landing sequence which will make the ship immeditaly take off / land again. That overlap could/is caused by the delay between the two tasks being generated.

Those Tasks are generated by the ScheduleShuttle task which is created at ShuttleZoneComponent::notifyInsertToZone.
So those duplicates are from tow calls to notifyInsertToZone. I picked a Shuttle OID on my server and added a raise(SIGTRAP) whenever that one is being loaded and made backtraces each time that is fired. That produced the two backtraces [1] and [2].

SceneObjects cause a transferObject() in their notifyLoadFromDatabase. transferObject causes the notifyInsertToZone calls and therefor a new task to be spawned.

But why is it being done twice? Probably because the ShuttleZoneComponent is part of a Vector (as a child of another object?!?).

This fix now restricts each shuttle to only have one task in the shuttleMap at a time.

The big question for me is now: Is it intended and deemd NOT harmful if an object gets "inserted" into the same zone twice (and also signaled, even thought it didn't "move") or is signaled as loaded twice?

[1] https://gist.githubusercontent.com/andir/fe9dd2b4ef3cd64b7aee/raw/5ceec3ff43ab1a662a4200ed2854f14edac4081e/gistfile1.txt
[2] https://gist.githubusercontent.com/andir/07a3e305eacf133d0699/raw/0620ce4ad0399db5ed45106b4c87a79a40cebb76/gistfile1.txt

Change-Id: Ie334f3523e161382ea1d7556904844ab5cb889e9